requirements.txt
django
roboflow
python-decouple